+++ src/main/java/com/example/taskflow/domain/auth/dto/request/LoginRequest.java
@@ -1,12 +1,8 @@
-package com.example.taskflow.domain.user.dto.request;
+package com.example.taskflow.domain.auth.dto.request;
 
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
