+++ src/main/java/com/example/taskflow/domain/auth/dto/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.example.taskflow.domain.user.dto.request;
+package com.example.taskflow.domain.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -6,7 +6,6 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
