+++ src/main/java/com/example/taskflow/global/config/FilterConfig.java
@@ -1,4 +1,4 @@
-package com.example.taskflow.config;
+package com.example.taskflow.global.config;
 
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
